requirements.txt
django
djangorestframework
requests
django-cors-headers
channels
pytorch
mediapipe
numpy
opencv-python